requirements.txt
absl-py==1.0.0
aenum==3.1.8
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.7.7
antlr4-python3-runtime==4.8
appdirs==1.4.4
argcomplete==2.0.0
arrow==1.2.2
astroid==2.11.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atari-py==0.2.9
atpublic==3.0.1
attrdict==2.0.1
attrs==21.4.0
autoflake==1.4
autopage==0.5.0
autopep8==1.6.0
backcall==0.2.0
bittrex-websocket-aio==0.0.0.3.0
black==22.3.0
botorch==0.6.3.1
Box2D==2.3.2
Box2D-kengz==2.3.3
box2d-py==2.3.8
bravado==11.0.3
bravado-core==5.17.0
cached-property==1.5.2
cachetools==5.0.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmake==3.22.3
cmd2==2.4.0
codetiming==1.3.0
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configobj==5.0.6
context-var==0.1.1
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
DateTime==4.4
DateTimeRange==1.2.0
decorator==5.1.1
Deprecated==1.2.13
devtools==0.8.0
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docker==5.0.3
dpath==2.0.6
dulwich==0.20.35
EasyProcess==1.1
entrypoints==0.4
executing==0.8.3
fasteners==0.17.3
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.31.2
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
gorilla==0.4.0
gpytorch==1.6.0
grandalf==0.7
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
gviz-api==1.10.0
gym==0.21.0
-e git+ssh://git@github.com/simondlevy/gym-copter.git@7236769b7586b92026d4b47f12363258c84d9508#egg=gym_copter
gym-notices==0.0.6
h5py==3.6.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
ipython==8.2.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
isort==5.10.1
iterextras==0.2.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonpatch==1.32
jsonpath-ng==1.5.3
jsonpointer==2.2
jsonref==0.2
jsonschema==4.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libcst==0.4.1
line-profiler==3.5.1
mailchecker==4.1.14
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mbstrdecoder==1.1.0
mccabe==0.7.0
memory-profiler==0.60.0
monotonic==1.6
more-itertools==8.12.0
msgpack==1.0.3
msgpack-python==0.5.6
msrest==0.6.21
multidict==6.0.2
multipledispatch==0.6.0
mypy-extensions==0.4.3
nanotime==0.5.2
networkx==2.7.1
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pexpect==4.8.0
phonenumbers==8.12.46
pickleshare==0.7.5
Pillow==9.1.0
pip-check-reqs==2.3.2
pipdeptree==2.2.1
platformdirs==2.5.1
plotly==5.6.0
pluggy==1.0.0
ply==3.11
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psrecord==1.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.11
py2puml==0.5.4
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
pyflakes==2.4.0
pygame==2.1.2
pygit2==1.9.1
pyglet==1.5.23
Pygments==2.11.2
pygtrie==2.4.2
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.7
pyperclip==1.8.2
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.18.1
pytest==7.1.1
python-benedict==0.25.0
python-dateutil==2.8.2
python-editor==1.0.4
python-fsutil==0.6.0
python-slugify==6.1.1
pytz==2022.1
PyVirtualDisplay==3.0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
redis==4.2.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==12.1.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schema==0.7.5
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
-e git+ssh://git@github.com/1057405bcltd/sentient-gym.git@24d3473316ab2472fce752cff9895a7b33567f11#egg=sentient_gym
sentient-traders @ git+ssh://git@github.com/1057405bcltd/sentient-traders.git@a784c4eacbea108f7f7cbfda3d24a5ae899697ce
-e git+ssh://git@github.com/1057405bcltd/sentient-util.git@664253f5f9f512b390116885f901a12268f03ae0#egg=sentient_util
setuptools-scm==6.4.2
shortuuid==1.0.8
shtab==1.5.3
signalr-client-aio==0.0.1.6.2
simplejson==3.17.6
six==1.16.0
slack-sdk==3.15.2
smmap==5.0.0
snakeviz==2.1.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.34
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.5.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.0
torchfile==0.1.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typed-ast==1.5.2
typeguard==2.13.3
typepy==1.3.0
typing-inspect==0.7.1
typing_extensions==4.1.1
uri-template==1.2.0
urllib3==1.26.9
userpath==1.8.0
uuid==1.30
validators==0.18.2
visdom==0.1.8.9
voluptuous==0.13.0
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==1.3.2
websockets==10.2
Werkzeug==2.1.1
wrapt==1.14.0
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.0
zope.interface==5.4.0
